src/main/java/org/fasttrackit/Person.java
package org.fasttrackit;

public class Person {
    private Integer id;
    private String name;
    private int age;

    public Person() {

    }

    public Person(String name, int age) {
        setName(name);
        setAge(age);
    }

    public Integer getId() {
        return id;
    }

    public String getName() {
        return name;
    }

    public int getAge() {
        return age;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public void setName(String name) {
        if (name == null) {
            throw new IllegalArgumentException("Name cannot be null");
        }
        this.name = name;
    }

    public void setAge(int age) {
        if (age < 0) {
            throw new IllegalArgumentException("Age cannot be negative");
        }
        this.age = age;
    }
}